[REF] Create custom hook for schema validation
Also move data dictionary hook into hooks
module
